+++ server/src-test/Main.hs
@@ -37,6 +37,7 @@ import qualified Hasura.IncrementalSpec             as IncrementalSpec
 -- import qualified Hasura.RQL.MetadataSpec      as MetadataSpec
 import qualified Hasura.CacheBoundedSpec            as CacheBoundedSpec
 import qualified Hasura.RQL.Types.EndpointSpec      as EndpointSpec
+import qualified Hasura.SQL.WKTSpec                 as WKTSpec
 import qualified Hasura.Server.AuthSpec             as AuthSpec
 import qualified Hasura.Server.MigrateSpec          as MigrateSpec
 import qualified Hasura.Server.TelemetrySpec        as TelemetrySpec
@@ -74,6 +75,7 @@ unitSpecs = do
   describe "Hasura.Server.Auth" AuthSpec.spec
   describe "Hasura.Cache.Bounded" CacheBoundedSpec.spec
   describe "Hasura.RQL.Types.Endpoint" EndpointSpec.spec
+  describe "Hasura.SQL.WKT" WKTSpec.spec
 
 buildPostgresSpecs :: HasVersion => Maybe URLTemplate -> IO Spec
 buildPostgresSpecs maybeUrlTemplate = do